requirements.txt
discord.py
pyyaml
httpx
flask
black
